requirements.txt
pyqt5
pandas
numpy
openpyxl
pyinstaller
msoffcrypto-tool
xlrd
